pyproject.toml
[tool.poetry]
name = "hw-11"
version = "0.1.0"
description = ""
authors = ["BombingArt <yaroshenkoyarik987@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
SQLAlchemy = "^2.0.31"
fastapi = "^0.111.0"
psycopg2 = "^2.9.9"
alembic = "^1.13.1"
uvicorn = {extras = ["standard"], version = "^0.30.1"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
fastapi-mail = "^1.4.1"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.3.4"
fastapi-limiter = "^0.1.6"
redis = "^5.0.7"
cloudinary = "^1.40.0"
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"


[tool.poetry.group.dev.dependencies]
sphinx = "^8.0.2"


[tool.poetry.group.test.dependencies]
httpx = "^0.27.2"

[tool.pytest.ini_options]
pythonpath = ["."]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

